Fix issues so that Python480 actually runs
Fixed issue with wrong variable type in spi write function.
Also added some commented code to toggle status LED based on Python480 output.

TODO: Start working on DMA for parallel bus to sdcard. Also need to verify sdcard communication is working

+++ v4_1/Atmel/Miniscope-v4-wire-free/Miniscope-v4-wire-free/python480.h
@@ -60,7 +60,7 @@ void python480SetFPS(uint32_t value)
 	}
 }
 
-void spi_BB_Write(uint8_t address, uint16_t value) 
+void spi_BB_Write(uint16_t address, uint16_t value) 
 {
 	int8_t i = 0;
 	//Bitbang SPI Master.
@@ -112,7 +112,7 @@ void spi_BB_Write(uint8_t address, uint16_t value)
 	
 }
 
-uint16_t spi_BB_Read(uint8_t address)
+uint16_t spi_BB_Read(uint16_t address)
 {
 	int8_t i = 0;
 	uint16_t value = 0;
